CD-408 Fixes handling of analysis criteria
It seems like transfer object with properties of type 'array' might return different types depending on usage. When nothing gets added using the 'add*' method of the 'array' field it will return an ArrayObject, a simple array will be returned if the 'add*' method got used. The generation of transfer object might need to get fixed to have a streamlined experience.